Expand collections and execution coverage
Add i64/u64 support across lowering and native emitters.

Expand collection/execution parsing and dump tests.

Update spec notes for current behavior.

+++ include/primec/Vm.h
@@ -7,8 +7,8 @@
 namespace primec {
 
 class Vm {
- public:
-  bool execute(const IrModule &module, int32_t &result, std::string &error) const;
+public:
+  bool execute(const IrModule &module, uint64_t &result, std::string &error) const;
 };
 
 } // namespace primec
